requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
argcomplete==1.10.0
    # via textract
beautifulsoup4==4.8.0
    # via textract
chardet==3.0.4
    # via
    #   -r requirements.in
    #   textract
docx2txt==0.8
    # via textract
ebooklib==0.17.1
    # via textract
extract-msg==0.23.1
    # via textract
imapclient==2.1.0
    # via extract-msg
lxml==4.6.3
    # via
    #   ebooklib
    #   python-pptx
numpy==1.20.3
    # via pandas
olefile==0.46
    # via extract-msg
pandas==1.2.4
    # via -r requirements.in
pdfminer.six==20181108
    # via textract
pillow==8.2.0
    # via python-pptx
pycryptodome==3.10.1
    # via pdfminer.six
pyenchant==3.2.0
    # via -r requirements.in
python-dateutil==2.8.1
    # via pandas
python-pptx==0.6.18
    # via textract
pytz==2021.1
    # via pandas
six==1.12.0
    # via
    #   ebooklib
    #   pdfminer.six
    #   python-dateutil
    #   textract
sortedcontainers==2.3.0
    # via pdfminer.six
soupsieve==2.2.1
    # via beautifulsoup4
speechrecognition==3.8.1
    # via textract
textract==1.6.3
    # via -r requirements.in
tqdm==4.60.0
    # via -r requirements.in
tzlocal==1.5.1
    # via extract-msg
xlrd==1.2.0
    # via textract
xlsxwriter==1.4.3
    # via python-pptx
